added inclusivity colorscheme
