+++ python_search/apps/terminal.py
@@ -47,13 +47,9 @@ class Terminal:
         return final_cmd
 
     def get_background_color(self):
-        if self.configuration.get_fzf_theme() == "light":
-            return " -o background=#ffffff "
 
         if self.configuration.get_fzf_theme() in ["dracula", "dark"]:
             return " -o background=#282a36 -o foreground=#ffffff "
 
-        if self.configuration.get_fzf_theme() in ["solarized"]:
-            return " -o background=#002b36 -o foreground=#ffffff "
-
-        return " "
+        theme = self.configuration.get_theme_object()
+        return f" -o background={theme.bg} -o foreground={theme.fg} "
